+++ examples/agdk-mainloop/src/lib.rs
@@ -1,4 +1,7 @@
-use android_activity::{AndroidApp, InputStatus, MainEvent, PollEvent};
+use android_activity::{
+    input::{InputEvent, KeyAction, KeyEvent, KeyMapChar, MotionAction},
+    AndroidApp, InputStatus, MainEvent, PollEvent,
+};
 use log::info;
 
 #[no_mangle]
@@ -8,6 +11,8 @@ fn android_main(app: AndroidApp) {
     let mut quit = false;
     let mut redraw_pending = true;
     let mut native_window: Option<ndk::native_window::NativeWindow> = None;
+
+    let mut combining_accent = None;
 
     while !quit {
         app.poll_events(
@@ -68,11 +73,56 @@ fn android_main(app: AndroidApp) {
                     if let Some(native_window) = &native_window {
                         redraw_pending = false;
 
-                        // Handle input
-                        app.input_events(|event| {
-                            info!("Input Event: {event:?}");
-                            InputStatus::Unhandled
-                        });
+                        // Handle input, via a lending iterator
+                        match app.input_events_iter() {
+                            Ok(mut iter) => loop {
+                                info!("Checking for next input event...");
+                                if !iter.next(|event| {
+                                    match event {
+                                        InputEvent::KeyEvent(key_event) => {
+                                            let combined_key_char = character_map_and_combine_key(
+                                                &app,
+                                                key_event,
+                                                &mut combining_accent,
+                                            );
+                                            info!("KeyEvent: combined key: {combined_key_char:?}")
+                                        }
+                                        InputEvent::MotionEvent(motion_event) => {
+                                            println!("action = {:?}", motion_event.action());
+                                            match motion_event.action() {
+                                                MotionAction::Up => {
+                                                    let pointer = motion_event.pointer_index();
+                                                    let pointer =
+                                                        motion_event.pointer_at_index(pointer);
+                                                    let x = pointer.x();
+                                                    let y = pointer.y();
+
+                                                    println!("POINTER UP {x}, {y}");
+                                                    if x < 200.0 && y < 200.0 {
+                                                        println!("Requesting to show keyboard");
+                                                        app.show_soft_input(true);
+                                                    }
+                                                }
+                                                _ => {}
+                                            }
+                                        }
+                                        InputEvent::TextEvent(state) => {
+                                            info!("Input Method State: {state:?}");
+                                        }
+                                        _ => {}
+                                    }
+
+                                    info!("Input Event: {event:?}");
+                                    InputStatus::Unhandled
+                                }) {
+                                    info!("No more input available");
+                                    break;
+                                }
+                            },
+                            Err(err) => {
+                                log::error!("Failed to get input events iterator: {err:?}");
+                            }
+                        }
 
                         info!("Render...");
                         dummy_render(native_window);
@@ -80,6 +130,73 @@ fn android_main(app: AndroidApp) {
                 }
             },
         );
+    }
+}
+
+/// Tries to map the `key_event` to a `KeyMapChar` containing a unicode character or dead key accent
+///
+/// This shows how to take a `KeyEvent` and look up its corresponding `KeyCharacterMap` and
+/// use that to try and map the `key_code` + `meta_state` to a unicode character or a
+/// dead key that be combined with the next key press.
+fn character_map_and_combine_key(
+    app: &AndroidApp,
+    key_event: &KeyEvent,
+    combining_accent: &mut Option<char>,
+) -> Option<KeyMapChar> {
+    let device_id = key_event.device_id();
+
+    let key_map = match app.device_key_character_map(device_id) {
+        Ok(key_map) => key_map,
+        Err(err) => {
+            log::error!("Failed to look up `KeyCharacterMap` for device {device_id}: {err:?}");
+            return None;
+        }
+    };
+
+    match key_map.get(key_event.key_code(), key_event.meta_state()) {
+        Ok(KeyMapChar::Unicode(unicode)) => {
+            // Only do dead key combining on key down
+            if key_event.action() == KeyAction::Down {
+                let combined_unicode = if let Some(accent) = combining_accent {
+                    match key_map.get_dead_char(*accent, unicode) {
+                        Ok(Some(key)) => {
+                            info!("KeyEvent: Combined '{unicode}' with accent '{accent}' to give '{key}'");
+                            Some(key)
+                        }
+                        Ok(None) => None,
+                        Err(err) => {
+                            log::error!("KeyEvent: Failed to combine 'dead key' accent '{accent}' with '{unicode}': {err:?}");
+                            None
+                        }
+                    }
+                } else {
+                    info!("KeyEvent: Pressed '{unicode}'");
+                    Some(unicode)
+                };
+                *combining_accent = None;
+                combined_unicode.map(|unicode| KeyMapChar::Unicode(unicode))
+            } else {
+                Some(KeyMapChar::Unicode(unicode))
+            }
+        }
+        Ok(KeyMapChar::CombiningAccent(accent)) => {
+            if key_event.action() == KeyAction::Down {
+                info!("KeyEvent: Pressed 'dead key' combining accent '{accent}'");
+                *combining_accent = Some(accent);
+            }
+            Some(KeyMapChar::CombiningAccent(accent))
+        }
+        Ok(KeyMapChar::None) => {
+            // Leave any combining_accent state in tact (seems to match how other
+            // Android apps work)
+            info!("KeyEvent: Pressed non-unicode key");
+            None
+        }
+        Err(err) => {
+            log::error!("KeyEvent: Failed to get key map character: {err:?}");
+            *combining_accent = None;
+            None
+        }
     }
 }